# docs/openapi/api/payment.yaml
openapi: 3.0.3
info:
  title: PayBuddy Payment API
  version: 1.0.0
  description: |
    PayBuddy 결제 도메인 API

    - 단일/복합 결제
    - 정기결제 (빌링키)
    - 결제 취소 및 환불
    - 차지백 및 분쟁 처리
    - Webhook 이벤트 수신
  contact:
    name: PayBuddy API Support
    email: api@paybuddy.com
servers:
  - url: https://api.paybuddy.com/v1
    description: Production
  - url: https://sandbox-api.paybuddy.com/v1
    description: Sandbox
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /checkout/submit:
    post:
      tags:
        - Payments
      summary: 결제수단 제출 및 저장
      description: |-
        결제수단을 제출하고 저장합니다. 서버가 인증(3DS, ISP) 필요 여부를 판단하여 NextAction을 반환합니다.

        **프로세스**:
        1. 결제수단 정보를 암호화하여 저장 (10분간 유효)
        2. 카드 유효성 검증 (번호, 만료일, CVC 등)
        3. 사기 탐지 및 한도 확인
        4. 인증 필요 여부 판단 (금액, 카드 발급국가 등)
        5. 인증 필요 시 인증 정보 반환, 불필요 시 PENDING_CONFIRM 상태 반환

        **인증 프로세스**:
        - 국내 카드 + 고액: ISP 인증
        - 해외 카드: 3DS 인증
        - 국내 카드 + 저액: 인증 불필요

        **다음 단계**:
        - 인증 필요: 클라이언트가 인증 페이지로 리다이렉트 → 인증 완료 후 confirm API 호출
        - 인증 불필요: 바로 confirm API 호출하여 최종 승인
      operationId: submitPaymentMethod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../schemas/payment.yaml#/PaymentSubmitRequest'
            examples:
              card_payment:
                summary: 카드 결제
                value:
                  payment_key: "01JGXM9K3V7N2P8Q4R5S6T7U8V"
                  payment_method_type: "CARD"
                  card:
                    card_number: "4111111111111111"
                    expiry_year: "25"
                    expiry_month: "12"
                    cvc: "123"
                    installment_months: 0
      responses:
        '200':
          description: 제출 성공, 인증 필요 여부 및 정보 반환
          content:
            application/json:
              schema:
                $ref: '../schemas/payment.yaml#/PaymentSubmitResponse'
              examples:
                authentication_required:
                  summary: 인증 필요 (3DS)
                  value:
                    payment_key: "01JGXM9K3V7N2P8Q4R5S6T7U8V"
                    status: "AUTHENTICATION_REQUIRED"
                    authentication:
                      type: "THREE_DS"
                      method: "POST"
                      url: "https://acs.card.com/auth"
                      data:
                        creq: "eyJ..."
                        termUrl: "https://pg.com/callback/3ds"
                authentication_not_required:
                  summary: 인증 불필요
                  value:
                    payment_key: "01JGXM9K3V7N2P8Q4R5S6T7U8V"
                    status: "PENDING_CONFIRM"
                    redirect_url: "https://merchant.example.com/pay/success"
        '400':
          description: 잘못된 요청
          content:
            application/problem+json:
              schema:
                $ref: '../error.yaml#/ErrorResponse'
        '401':
          description: 인증 실패
          content:
            application/problem+json:
              schema:
                $ref: '../error.yaml#/ErrorResponse'
        '422':
          description: 처리할 수 없는 요청 (카드 검증 실패 등)
          content:
            application/problem+json:
              schema:
                $ref: '../error.yaml#/ErrorResponse'
              examples:
                card_validation_failed:
                  summary: 카드 검증 실패
                  value:
                    type: 'about:blank'
                    title: 'Payment validation failed'
                    status: 422
                    detail: '허용되지 않는 카드 브랜드입니다'
                    error_code: 'CARD_BRAND_NOT_ALLOWED'
  /payments/confirm:
    post:
      tags:
        - Payments
      summary: 결제 진행 (승인/매입확정 통합)
      description: |-
        결제를 진행합니다. 서버가 현재 상태(status)에 따라 동작을 결정합니다.

        - status=INITIALIZED 또는 PENDING: 승인(Authorize)을 수행
          - 서버 정책에 따라:
            - AUTHORIZED: 가승인(Preauthorize) 완료 (추후 재호출로 매입확정)
            - CAPTURED: 승인과 동시에 매입 확정(즉시 결제)
        - status=AUTHORIZED: 매입 확정(Capture)을 수행 → CAPTURED

        클라이언트는 void/cancel처럼 세부 동작을 선택할 필요가 없으며,
        동일 엔드포인트 재호출로 흐름을 진행합니다.
      operationId: confirmPayment
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '../schemas/payment.yaml#/PaymentConfirmRequest'
      responses:
        '200':
          description: 진행 결과
          content:
            application/json:
              schema:
                $ref: '../schemas/payment.yaml#/PaymentConfirmResponse'
        '400':
          description: 잘못된 요청
          content:
            application/problem+json:
              schema:
                $ref: '../error.yaml#/ErrorResponse'
        '401':
          description: 인증 실패
          content:
            application/problem+json:
              schema:
                $ref: '../error.yaml#/ErrorResponse'
        '422':
          description: 처리할 수 없는 요청
          content:
            application/problem+json:
              schema:
                $ref: '../error.yaml#/ErrorResponse'
  /payments/{payment_key}:
    get:
      tags:
        - Payments
      summary: 결제 상세 조회 (운영/관리자용)
      description: |-
        가맹점 및 운영자가 결제 상세 정보를 조회합니다.
        - 민감정보 포함 (수수료, 정산금액 등)
        - API Key 인증 필요
      operationId: getPayment
      parameters:
        - name: payment_key
          in: path
          required: true
          schema:
            type: string
          description: 결제 고유키
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '../schemas/payment.yaml#/PaymentDetailResponse'
        '404':
          description: 리소스를 찾을 수 없음
          content:
            application/problem+json:
              schema:
                $ref: '../error.yaml#/ErrorResponse'
  /payments/{payment_key}/receipt:
    get:
      tags:
        - Payments
      summary: 결제 영수증 조회 (소비자용)
      description: |-
        소비자가 결제 영수증을 조회합니다.
        - 민감정보 최소화 (수수료 제외, 카드번호 강한 마스킹)
        - 공개 가능 (토큰 기반 인증)
        - PDF/HTML 형식 지원
      operationId: getPaymentReceipt
      parameters:
        - name: payment_key
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum:
              - json
              - pdf
              - html
            default: json
          description: 응답 형식
      responses:
        '200':
          description: 영수증 조회 성공
          content:
            application/json:
              schema:
                $ref: '../schemas/payment.yaml#/ReceiptResponse'
            application/pdf:
              schema:
                type: string
                format: binary
            text/html:
              schema:
                type: string
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 가맹점 API Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 토큰 (내부 서비스용)